+++ src/gufo/thor/config.py
@@ -156,6 +156,8 @@ class ExposeConfig(object):
         web: Web listener configuration.
         port: An HTTPS port of the NOC's user interface (deprecated).
         open_browser: Open browser on startup.
+        mtls_ca_cert: When set, enables mTLS and defines CA certificate path,
+            relative to `etc/envoy`
     """
 
     domain_name: str = "go.getnoc.com"
@@ -163,6 +165,7 @@ class ExposeConfig(object):
     mongo: Optional[Listen] = None
     postgres: Optional[Listen] = None
     open_browser: bool = True
+    mtls_ca_cert: Optional[str] = None
 
     @staticmethod
     def from_dict(data: Dict[str, Any]) -> "ExposeConfig":
@@ -189,6 +192,16 @@ class ExposeConfig(object):
             data["mongo"] = Listen.from_dict(data["mongo"])
         if "postgres" in data:
             data["postgres"] = Listen.from_dict(data["postgres"])
+        # mTLS
+        if "mtls_ca_cert" in data:
+            # Check
+            with errors.context("mtls_ca_cert"):
+                path = Path("etc", "envoy") / Path(data["mtls_ca_cert"])
+                if not path.exists():
+                    errors.error(
+                        f"File {path} is not found. "
+                        "Place proper CA certificate to enable mTLS"
+                    )
         # Deprecated port option
         if data.get("port"):
             port = data.pop("port")